Created css-in-js-helpers package
