+++ packages/vscode-css-in-js/package.json
@@ -22,7 +22,7 @@
         "vscode": "^1.13.0"
     },
     "scripts": {
-        "build": "babel --no-babelrc src -d dist",
+        "build": "babel src -d dist",
         "postinstall": "npm run build",
         "vscode:prepublish": "npm run build",
         "prettier": "prettier-standard 'src/**/*.js'",
@@ -71,9 +71,7 @@
         "vscode": "^1.0.0"
     },
     "dependencies": {
-        "css": "^2.2.1",
-        "stringify-object": "^3.2.0",
-        "stylis": "^3.2.3"
+        "css-in-js-helpers": "^1.0.0"
     },
     "icon": "icon.png",
     "galleryBanner": {
